+++ src/main/java/com/sportynote/server/repository/NoteNodeRepository.java
@@ -17,7 +17,7 @@ public class NoteNodeRepository{
     public void save(NoteNode node) {
         em.persist(node);
     }
-    public NoteNode findById(Integer nodeId){
+    public NoteNode findById(Long nodeId){
         return em.createQuery("select n from NoteNode n where n.idx =: nodeId", NoteNode.class)
                 .setParameter("nodeId", nodeId)
                 .getSingleResult();
