+++ src/main/java/com/sportynote/server/repository/query/RoutineMachineDto.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @AllArgsConstructor
 public class RoutineMachineDto {
     //내 기구 정보 내 루틴 idx
-    private Integer routineIdx;
-    private Integer machineIdx;
+    private Long routineIdx;
+    private Long machineIdx;
     private String machineName;
     private String targetArea;
     private String url;
